refactor: split game page into lobby and play page
