+++ app/game/[code]/play/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function PlayPage() {
+  return <div>wip</div>;
+}
